refactor: 拆分 interface: McpServerService
