+++ backend/sdk/src/main/java/com/alibaba/higress/sdk/model/authorization/CredentialTypeEnum.java
@@ -12,18 +12,61 @@
  */
 package com.alibaba.higress.sdk.model.authorization;
 
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+
+import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
+
+import lombok.Getter;
+
 /**
  * @author lvshui
  */
+@Getter
 public enum CredentialTypeEnum {
     /**
      * credential type is key-auth
      */
-    KEY_AUTH("key-auth");
+    KEY_AUTH("key-auth", "API_KEY");
 
     private final String type;
+    /**
+     * aliases for credential type. ignore case
+     */
+    private final List<String> typeAliases;
 
-    CredentialTypeEnum(String type) {
+    CredentialTypeEnum(String type, String... aliases) {
         this.type = type;
+        this.typeAliases = Arrays.asList(aliases);
+    }
+
+    private static final Map<String, CredentialTypeEnum> CACHE = new HashMap<>();
+    static {
+        for (CredentialTypeEnum value : CredentialTypeEnum.values()) {
+            CACHE.put(value.getType().toUpperCase(), value);
+
+            List<String> typeAliases = value.getTypeAliases();
+            if (CollectionUtils.isEmpty(typeAliases)) {
+                continue;
+            }
+            for (String typeAlias : typeAliases) {
+                CACHE.put(typeAlias.toUpperCase(), value);
+            }
+        }
+    }
+
+    public static CredentialTypeEnum fromType(String type) {
+        if (StringUtils.isBlank(type)) {
+            throw new IllegalArgumentException("type cannot be blank.");
+        }
+        CredentialTypeEnum result = CACHE.get(type.toUpperCase());
+        if (Objects.isNull(result)) {
+            throw new IllegalArgumentException("Unknown credential type: " + type);
+        }
+        return result;
     }
 }
