+++ backend/sdk/src/main/java/com/alibaba/higress/sdk/service/mcp/save/McpServerSaveStrategy.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2022-2024 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package com.alibaba.higress.sdk.service.mcp.save;
+
+import com.alibaba.higress.sdk.model.mcp.McpServer;
+
+public interface McpServerSaveStrategy {
+    /**
+     * support mcp server
+     *
+     * @param mcpServer mcp server save param
+     * @return true or false
+     */
+    boolean support(McpServer mcpServer);
+
+    /**
+     * add or update mcp server
+     *
+     * @param mcpServer mcp server
+     * @return mcp server
+     */
+    McpServer save(McpServer mcpServer);
+
+}
